requirements.txt
requests>=2.31.0
pyyaml>=6.0
jsonschema>=4.17.0
pytest>=7.4.0
prance>=23.6.21.0
openapi-spec-validator>=0.6.0
colorama>=0.4.6
jinja2>=3.1.2
